[interface] Add new DifferentiableFunctionInterface
The new interface is modeled by std::function and
is entirely based on value-semantics.

Our tests show that, because of return-value-optimization
there is no performance overhead when returning copies
instead of writing into given non-const references. We tested
this with Range=FieldMatrix<double,n,n> for n=2,3,50.

The tests indicate that sometimes returning the value is
even slightly faster.