[cleanup] Make GridViewEntitySet copyable
We need this in order to have value semantics of
grid functions. Hence I switched from storing a
GridView reference to a pointer.

This means GridViewEntitySet is only a thin wrapper
and you will always have to store the GridView youself.
You will often do this anyway, e.g. in a global FE basis.
This way we avoid having to store two GridView.

However, we may also want to store a copy instead.