[interface] Add entitySet() method
This provides access to the EntitySet of the wrapped
function. Currently I've implemented this as member
function as before. We may also want to change this
to a free function. However I suspect that this may
often lead to name conflicts avoiding that the
free function is found via ADL.

Having this added the new interface is on par with the old one.