Add DifferentiableFunctionFromCallables class
This wrapper implements the DifferentiableFunction interface
using a given list of callables. In contrast to the old
FunctionFromCallable we do not store std::function objects
in order to avoid the hidden type-erasure overhead.
This comes at the price of a more complicated type that
also incorporates the types of the callables statically.

However this can nicely be hidden by calling the
free function makeDifferentiableFunctionFromCallables
that passes the correct parameters. Notice that
use still have to pass the signature and derivative traits
via a tag class.