[bugfix] Remove DUNE_FINAL
This is not allowed if the base function is Dune::Function. This happens
to be the fact during interpolate() if the local FE basis is not derived
from the polymorphic base class.

Notice that this will not degrade performance of interpolate.
In case of the virtual local FE base class 'final' wil not help
and in the other case there's no virtual call anyway.