Make interpolate() more flexible.
Now the range of the function and the block type of
the coefficient vector need not be the same anymore.
We only need to have them the same number of entries.
E.g. you can now use FieldVector<K,n*m> coefficients
if the function range if FieldMatrix<K,m,n>.