Reduce default buffer size to 56
As a consequence the overall size of PolymorhicSmallObject
will be 64 (mind the additional pointer) and thus more likely
to fit in a cache line.

Notice that I did not do any performance testing on this.
But buffersize=64 was arbitrary anyway. Now there's at least
a weak argument for the default choice. Feel free to change
this if there are better arguments for another default size.