Implement local views and index sets properly
Previously, local views on the B-spline basis would return the restrictions
of _all_ basis functions to the given element.  This include all basis functions
that were zero on the element in question.  Of course this could only be a
temporary state.

This patch implements the proper selection and numbering of only the non-zero
basis functions for any given element.  So far, it only works for 1d, though.
The general case is quite a Gefummel, and I will do that in a separate commit.