Fix computation of the number of shape functions for a given element
The previous formula failed on very small grids.

Incidentally, with this fix the basis passes my complete test suite, which means
that constant-zero shape functions will be produced anymore even for not-1d-grids.

Note however that evaluation of shape function values and derivatives is still
linear in the number of knots per coordinate direction.  That hack will need to
go soon, too.