Add example solving the Stokes equation with Taylor-Hood elements
Here is a second example to show how our function space bases are used.
This one focuses on the Taylor-Hood element as a simple example for a
non-trivial finite element tree.  The code should implement the well-known
driven-cavity problem: unit square, no source term, homogeneneous Dirichlet
boundary conditions on three sides of the square, constant (non-zero)
Dirichlet bc on the forth side.

Unfortunately this example program is incomplete as of yet, because
we are lacking some infrastructure.
- The problem is set up but not actually solved, because apparently there
  is no appropriate solver for this kind of problem in dune-istl.
- In addition to the TaylorHoodBasis class we use the PQ1NodalBasis
  and PQ2NodalBasis classes: To interpolate, and to be able to add
  the result to a VTK file.