[test] Allow to disable parts of the local tests
We cannot generally expect from a scalar basis that its LocalFiniteElement
fulfills the entire interface.  For example, in a B-spline basis, the LocalCoefficients
functionality cannot be implemented, because B-spline basis functions cannot
uniquely be assigned to grid entities.  Therefore we allow that parts of the
tests are disabled for certain bases.  For this we simply use the infrastructure
already provided by test-localfe.hh from the dune-localfunctions module.