Implement the method BSplineLocalBasis::order
Unfortunately, the general interface of the LocalBasis class mandates that the 'order' method
takes no arguments, and returns a single integer.  It therefore cannot reflect that fact that
a B-spline basis function can easily have different orders in the different coordinate directions.
We therefore take the conservative choice and return the maximum over the orders of all directions.