Assign local dofs to grid subentities just as in the Qk case
This may seem strange at first: you cannot consistently do such an assignment for
the basis functions of a B-spline basis.  However, for the important case of
open knot vectors, this assignment allows to extract the dofs on the grid boundary.
And that is pretty helpful.

This implementation only works for 1d and 2d grids.  I still haven't found a general
rule to construct the assignment, and I have been too lazy to adjust the longish
3d case from QkLocalCoefficients yet.

Also, this may fail for the zero order case.  It wouldn't really make much sense
there anyway.

Plus: if it fails, it fails without warning, because otherwise you would get
a lot of warnings.  Note that failing coefficient assignement does not mean that
the BSplineBasis becomes unusable.  Only you won't be able to find the
boundary dofs.