Add tupletreepath.hh
This contains some functionality for using tuples as hybrid
static/dynamic multi-indices for a TypeTree. The main idea
is that each entry is either an int (dynamic) or an
integral_constant<int,i> (static). Then the standard
tuple machinery does almost everything we need.

I only added convenience methods for appending static/dynamic
indices, because this is what we do, when we build a tree
starting from its root.

Furthermore I implemented getChild(tree, treePath) returning
the child in 'tree' that is addressed by the tuple treePath.
As a consequence you can e.g. do

  getChild(tree, std::make_tuple(_0, 3, _1))

assuming that you defined _N = integral_constant<int,N>.
Maybe we can move this to dune-typetree.

As a side product this contains an Std::interger_sequence
implementation mimicing C++14 together with a helper producing
a static integer range. This allows to very easily implement
tuple_slices. The latter could be a candidate for dune-common.