[cleanup] Rename PQkNodalBasisLeafNode to PQkNode
This is no longer tied to the PQkNodalBasis, so 'NodalBasis'
should be remove. Regarding the 'Leaf' part:

Every basis-core^tm (reusable part of a global basis)
consist of a NodeFactory,NodeIndexSet,Node and we should
have a consistent naming scheme. Hence 'Leaf' must also
died (unless we have a TaylorHoodSecondLevelNode).