Adjust to new basis-core interface
The TaylorHoodBasis is now implemented on top of the
new basis-core concept, using the PQkNodeFactory and
not the PQkNodalBasis. Furthermore it already correctly
implements the slight changes in the global basis user
interface:
* It provides size() and as-static-as-possible treePath()
  in its nodes.
* Local indices are flat and lexicographic now and no longer
  interleaved (this should not influece _correct_ user code).

However, it's currently not implementing the basis-core
concept itself. I'll do the transition within the next
days. Notice, that this should not influece the user-part
of the interface.