Add support for nested indices
By using the HierarchicVectorBackend, we can now interpolate for
nontrivial ansatz trees if the global indices match the container
structure. However, interpolation is only done wrt a single leaf
node in the ansatz space, that has to be specified via a tree path.

By using the FlatIndexVectorBackend, one can also
(as before) interpolate into implicitly given power spaces
if the basis is scalar, but the vector entries are blocks.

This can even be combined to interpolate, e.g., wrt a vector
valued Taylor-Hood-space.

The current implementation has the following drawback:
Before the commit we skipped elements, where all DOFs
have already been interpolated. With the new version
this is deactivated until I find a nice way to derive
a proper bitfield type for storing this information

Notice that the 4-argument version of the interpolate()
method is not backward compatible, because you now have
to specify the tree path.