Allow to interpolate into nontrivial subtrees
Now you can, e.g., directly interpolate a given vector-field into
the velocity component of a Taylor-Hood-element. For this you have
to provide a NodeToRangeEntry map that specifies how leaf nodes
in the ansatz tree are mapped to entries in the range of the given
function.

I'm planning to extend DiscreteScalarGlobalBasisFunction accordingly,
such that you can use the same mechanism to define, e.g., a vector field
by a power node. Then we'd no longer have to copy entries around for
writing the grid function.

I'm planning to implement a default NodeToRangeEntra map based on the
lexicographic order.