Add member/getter/setter treeIndex to nodes
Currently one can only identify a node uniquely via its treepath.
If you want to associate data to nodes, this is very unhandy,
because you cannot put those treepaths in an associative container,
because their types differ. Furthermore the lexicographic node-index
(not the dof index) cannot be computed staring from the node alone.

As a remedy this commit introduces an additional node member
treeIndex which is the lexicographic index of the node. While
these are in principle const and could be set during tree
construction, this would have to be done manually in every
node. With the implemented solution these indices are computed
and set by the treeBind method, such that node implementations
do not have to be modified.

If we'd like to keep these members, we may want to set them
on construction to avoid setting the same indices again on
every bind.