Add SizeInfo
This class encapsulates size information of a basis.
I propose to return something like this on basis.size()
without arguments.

If we decide to do so, the concrete type can be implementation
defined. Then we can still use this class as default implementation.
Until then this can still be used as drop in replacement for the
basis inself when calling HierarchicVectorWrapper::resize().