Only create HierarchicVectorWrapper if needed
We now support to directly pass a vector having suitable multiindex
access.

This call may also be removed completely. Than one must always pass
a vector type that provides multiindex access. This however is given
for plain vectors if the multiindex is flat. So a wrapper must only
be created in the complicated cases.