Fix size methods
If the user wants to query the multiindex size recursively
by giving prefixes, he must be able to terminate the recursion.
Hence he must know, which multiindex is a leaf in the index tree.
This can easily be achived by guaranteeing, that the user
can grow the prefix until the first zero is obtained.