Terminate HierarchicalVectorWrapper- resize recursion if size(prefix)=0
If the size(prefix) for a given prefix is zero the
index corresponds to a leaf in the index tree, i.e.,
a single dof. If the vector entry for this prefix
is dynamically sized we have to consider two cases:

a)It's size is already nonzero. Then we must assume that
  the user already resized it to whatever size he wants.
  Resizing it to one would destroy the possibility
  to have implicit power spaces with runtime
  power.

a)It's size is zero. Then we must assume that the user
  wants to use an implicit power space, but we cannot
  know the power. Hence we throw an exception.

In both cases we can terminate the resize recursion.