Don't stop resize recursion for statically sized vector entries
It's not unlikely to have vectors of the form array<vector<Foo>>
where we cannot stop to traverse the nesting hierarchy further
to resize the inner dynamically sized vectors.

This comes at the price of looping over all entries of a dynamically
sized vector and checking all subentries. Hopefully the compiler is
smart enough to see that we don't do anything because all the checks
are known to be satisfied at compile time.

To get rid of this we'd need more information on the index structure.
Something like indexIsUniform(prefix) would be sufficient here.
It should return true only if if the index subtree does always
look the same for all prefixes

  (prefix,0)...(prefix,size(prefix)-1)