Add support for truely hybrid multi-type containers
We only require that there is a common coeffient type CO
such that the individual coeffients associated to single
global basis functions can be cast to CO. Such a coefficient
for a basis function is the result of vector[i1][i2]...[in]
where (i1,i2,...,in) is the multi-index of this basis function.

Since the multi-index may be dynamically sized and non-uniform
in case of mutli-type containers, one must pass the type of CO
in order to provide the compiler a termination criterion for
the recursive function instantiation.