Get rid of old fixed-multiindex-size implementation
Now, if the coefficient type 'void' is given and the MultiIndex
has static size, the result of applying MultiIndex.size()-times
[_0] to the vector type. This resembles the old behaviour.

This is currently done in the operator[] method, because only
this one knowns the MultiIndex type. I have the impression
that its cleaner to make the MultiIndex a template parameter
of the whole class. However, you would then have to pass more
types to the wrapper.