Add some candy to simplify construction of basis
Inspired by Christians talk you can now use the following
snipped to construcct a Taylor-Hood basis from scratch.
Still a little clumsy but I'm afraid that's as easy as
we can get:

  using namespace Functions::BasisTags;     // For the tags determining index blocking
  using namespace Functions::BasisBuilder;  // This is where power,composite,pq, makeBasis are defined

  auto basis = makeBasis<MultiIndex>(       // Make a basis out of ...
    gridView,                               // ... a grid view
    composite(                              // ... and a composite node consisting of
      power<dim>(                           //      ... a power node with dim
        pq<2>(),                            //          ... PQ2 nodes
        BlockBackTag()),                    //          ... where the power node indices are blocked via an appended index
      pq<1>()                               //      ... and a PQ1 node
    ));

To add your own node-factory to the machinary you have to
create a builder encapsulating the basis-specific data
and a free function to simplify creation of the builder.
See the PQkNodalBasis for an example.