Add makeBasis with default MultiIndex size
This can be computed from the requiredMultiIndexSize
exported by the factory builder. Notice that we can't
derive it from the factory itself because we want to delay
its instanciation until the user provides a grid view.
Hence the factory builder (which is invariant under
the grid view) has to export it.

I'm a little surprised that these two templates work.
Depending on wether you use

  makeBase(...) or makeBases<MultiIndex>(...)

the compiler indeed selects the correct one.
Maybe someone can have a look if this is OK
or if it's gcc beeing to gracefull.