Use a manually build Taylor-Hood basis
This is to demonstrate how the basis selection mechanism works.
Using the explicit TaylorHoodBasis still works. However we now
have to explicitly provide the coefficient type, i.e., the
type used for single dof coefficients. This is because the
generically constructed basis used a dynamically sized MultiIndex
such that the coefficient type can lonoger be derived automatically.