Add ReservedDeque
Beeing very similar to reserved vector, this provides a deque
with static capacity. Using this as replacement for ReservedVector
in SizePrefix or as MultiIndex would allow to avoid some index
shuffling. But I'm undecided if this is really better. Just
added this for review. Maybe this will be removed again.