Add helper template alias that allows to constrain constructors
If a class Foo has a constructor with a forwarding reference,
one has to be very carefull to avoid that this one is used
when copying a tuple containing a Foo. If all arguments are just
forwarded to a contained member of type Bar, this can be avoided
by only providing the Foo constructor if there's a corresponding
Bar constructor. This template alias simplifies the necessary
enable_if trickary.

Thanks to Felix Gruber for pointing me to this problem!