Remove virtual size() method
Instead we now rely on the size() method in the LeafBasisNode
base class which requires that we call setSize() during bind.

The latter could in principle be done automatically by the
BindVisitor by calling finiteElement().size(). But there may
be cases where constructing the FE is expensive such that
one would do this on demand only, while the size is apriorily
known. If you feel this is overly pessimistic feel free to add
this call to the leaf method of the BindVisitor.