Add ExpandTuple
For given T and tuple<Args...> this is an alias to T<Args...>.

If you want to propagate a variadic template argument list
you have the problem that you can't export the unexpanded
list. Instead you can only export the expanded list, e.g.,
encapsulated in a tuple.

For such a tuple ExpandTuple allows to re-expand the
list into another tuple.