[concept] Use const& for concept checks
If we check for an interface concept there's no need
to usie perfect forwarding. In fact it may polute the
type by &&. Furthermore using const& is much more
expressive since it allows to check for constness
of methods. So here's the guideline for interface
concepts:

  If you want to check if a type models a interface
  concept, always use 'const T&'. Then check the const
  methods directly and the non-const methods using
  a static cast.

We could enforce this in the 'models' method, however
it may still be usefull to allow && in special situations
where you want to check for special features of the type.