[cleanup][concept] Clean up concept-check implementation
Now all (but one) of the internal concept check
methods return a std::true_type/std::false_type
instead of a plain bool. This unifies the implemenation.

Furthermore in the central check if the expressions
defining a concept are valid the ugly void* vs. T*
hack for selecting the proper overload is replaced
by using PriorityTag<...> which makes this easier to understand.