[conecpt][bugfix] Fix bug in modelsConceptList
Before this checking with multiple refined concepts failed
because the recursion for modelsConceptList does not work.
The reason for this is, that the declaration is not complete
during the definition of the trailing return type. Hence
we can't use the same function in the trailing return type
recursively.

As a solution we can go back to return constexpr bool instead
of true_type/false_type. Then the recursion happens in the definition
and not in the declaration.

This commit also simplifies the check for refined concepts,
by dropping the RefinedConcept base class and friends. Instead
we now directly check if BaseConceptList is present.

Furthermore this commit drops support for checking a typelist
of several concepts at once via models. This was never used and
should be done in a separate function if we really need it.

Now the whole concept checking boils down to seven easy to
understand one-line methods.