Workaround an issue with clang
It seems that clang has a problem with either 'const auto&'
or 'auto&&' here. Because this is just used for an
integral_constant there's no penalty in using plain
'auto' here.

Thanks to Elias Pipping for pointing out the problem.

Notice that this seems to be related to this issue:

http://stackoverflow.com/questions/33872039/invalid-explicitly-specified-argument-in-clang-but-successful-compilation-in-gcc
https://llvm.org/bugs/show_bug.cgi?id=25693
(cherry picked from commit 1bd66e601e9ebb982c3e424e38a2e07fa9af922f)