[cmake][bugfix] Correct handling of the EXCLUDE_FROM_ALL property:
The signature of `dune_add_test`, that takes the `TARGET` from the user
produced excluded targets if added in a directory that has the directory
property `EXCLUDE_FROM_ALL` set. The correct solution IMO is to not use
that property on test subdirectories anymore (it is a relic from the old
testing magic), because we rely on `dune_add_test` to handle exclusion.

See core/dune-common#21 for details.