Add SubspaceBasis and SubspaceLocalView
These allow to encapsulate the subspace spanned by a subtree
associated to a prefix tree path. A subspace basis is not a global basis
itself, because its local and global indices, tree paths, and tree indices
contain holes because localIndexSet() gives you the local index set
of the root basis and tree() gives you a subtree of the original
root tree.

By the previous commit any global basis is also a subspace basis
which has itself as root basis and an empty prefix path.