Node base classes will now always use std::size_t
Notice that I kept a dummy template parameter in place of the
size_type parameter. If we remove this, we change the implementers
interface. This could simple be fixed in all basis-node implementaions
in dune-functions. But I'd like to avoid breaking user code for those
who implemented their own bases.