[doc][manual] Replace tensor product by product/external sum
The the constructions denoted 'sum' and 'tensor product'
before are actually the 'internal sum' and the 'external sum'
of spaces. In terms the sets the latter coincides with the
Cartesian product which is also simply denoted as product space.

The commit cleans up the terminalogy. It also replaces the missused
\otimes by \times or \oplus.

For the composed basis of the product space I'm not aware of any
standard terminology and simply called it concatenated basis using
the notation (B,C).