Allow to evaluate the derivative (gradient) of a local function.
For this purpose add a specialization for the FlatVectorBackend.

For a vector valued function constructed via a power basis the
returned gradient is a vector of vectors isomorphic to a square
matrix. The components [i][j] return the derivative of the i-th
component wtr. to the j-th variable.

NOTE: The data type for derivatives (JacobianRange) is wrong.
Comments are added. A dirty fix is added, which is only working
for scalar field and vector valued fields with dimension equal
to the geometric dimension.