Add algorithms for working with function space bases
This is a collection of algorithms working on function space bases.
It does e.g. allow to
* loop over all (local view, local index set) pairs for a given basis
* loop over all DOFs associated to an intersection
* loop over all boundary DOFs

This is not realized by fixed utility methods that do, e.g., mark
the DOFs. Instead these are stl-like algorithms invoke given callback
functions. This allows for greater flexibility and composability.

The namespace TreeAlgorithms contains some algorithms for looping
over a type tree. These are proposed for dune-typetree. However I
copied them here to simplify the implementation. They can be removed
once they have arrived in dune-typetree.

Furthermore the algorithms for looping over sub-sub-entities of a
sub-entity in a reference element may be moved to dune-geometry.