Use forEachNode() in bindTree()
This is just for demonstration. Unfortunately it only
works when making the non-const methods of BasisNodeMixIn
public because you cannot list lambdas as friends. However,
I believe that this is not a big problem, because we only
ever handout a const tree such that these functions are
hidden from the user anyway.