[DefaultLocalIndexSet] Cache indices on bind()
This patch changes the DefaultLocalIndexSet to cache all indices on bind.
Moreover, it introduces a new interface method
`NodeIndexSet::indices(Iterator)`.This methods takes a RandomAccessIterator and
is supposed to fill the underlying storage with its indices. If the node does
not implement that method, DefaultLocalIndexSet falls back to iterating over the
index range and calling `NodeIndexSet::index()` for each index.