Change concept `NodeIndexSet` to check for indices instead of index
Implementing `indices()` via `index()` will be very slow.
On the otherhand implementing `index()` via `indices()`
would be even worse. Hence we will not be able to have
both at the same time without code duplication. Since
this is an internal implementers interface we drop
`index()` in favour of `indices()`.

Notice that the `LocalIndexSet` still only provides
the convinient `index()` method. This is OK, because
it can(has to) fill an internal cache with the result
of `NodeIndexSet::indices()`.