Rename NodeFactory to PreBasis
We discussed the renaming on the last meeting. Reasons for renaming:
* NodeFactory does not fit because it creates more than just nodes
* The FactoryTag/FactoryBuilder classes are really factories that create
  a NodeFactory, but we would not like to se NodeFactoryFactory
* Pre-Basis is more understnadable: It's almost a basis.

I renamed all occurences in dune-functions. If you implemented
your own node-factory, you only need to take care that the concept
name was changed from NodeFactory to PreBasis. The actual naming
of your class is am implementation detail. However, for consistency
you may want to adapt this also.