Move Carsten's text on makeBasis and preBasis to the 'implementor' chap.
As agreed upon on the phone, the first part of it targets people
that want to understand how it works rather than people that
simply want to use it.

The second part somewhat doubles my own description of the makeBasis
mechanism (which is currently 2.6).  I don't see much sense in merging
the two.  But I don't just want to delete it, either.  So for the time
being it goes into the final chapter.