Use static assertion to ensure that a coefficient deduction
If HierarchicVectorView is not passed the coefficient type
directly, it will try to deduce it which fails for dynamically
sized multi-indices. This is e.g. the case for non flat indices
if you do not create the wrapper manually.

This problem is now captured by a static_assert. In order to see
this in the error message the default coefficient type used to indicate
that it should be deduced had to be changed from void to a named
type. Otherwise you get errors for code using the result before
even seeing the failed assertion error. Using a named type here
is a good idea anyway.