Speed up index computation for the first-order case
If the order k is 1, the compiler should be able to deduce that
only vertex dofs can ever appear.  However, the current implementation
of the 'indices' method did not use this fact.  This minor patch
adds a dedicated compiler hint for the k==1 case, which leads to a
measurable speedup.  See

  https://gitlab.dune-project.org/staging/dune-functions/issues/30