Fall back to istlVectorBackend() in interpolate()
If the passed vector does not satisfy the VectorBackend
interface, it should be wrapper. Before, the latter was
done using HierarchicVectorWrapper. This is replaced by
ISTLVectorBackend.

While ISTLVectorBackend does only support scalar coefficients,
that's no problem here, because the automatic wrapping into
HierarchicVectorWrapper was also restricted to this case.

Since I'm not convinced that a global makeVectorBackend()
function that wraps if needed and forwards otherwise
is necessary this is now done inline using a simple lambda.