Implement a persistent pseudo grid view
This caches any accessed index lazily. The grid
view only implements indices but no iteration
because this is not needed for the intended use:

Create a persistent version of a basis that
survives grid modifications in order to interpolate
solutions to a refined grid. Notice that this approach
is non-intrusive, i.e. the bases don't have to be adjusted,
but simply work, on the PersistentGridView, as long as they
only rely on the provided feature set.