Remove the deprecated GeometryType::basic_type
For the moment this involving always using the virtual interface of
dune-localfunctions. It  is a first step towards implementing
the Raviart-Thomas basis for grids with more than one element type,
but quite a bit more work is missing for that.
On the other hand, one grids with a single known element type the
patch may lead to a minor slowdown.  This will be fixed in a
subsequent patch.