[bugfix] Use PowerPreBasis::MultiIndex::max_size as size for SizePrefix
The SubPreBasis::SizePrefix::max_size may be to small,
if the current PowerPreBasis addds a digit. The solution
if to use PowerPreBasis::MultiIndex::max_size. This may
be suboptimal if this PowerPreBasis is an inner node
and more digits are added later. Before trying to optimaze this,
one should benchmark which solution is better:
* Use an optimal (as small as possible) SizePrefix
  in each node.
* Use the same SizePrefix for all nodes to allow
  modification insteaf of copying.