[doc][manual] Be more precise abot child() method
Only non-epmty paths are supported (for now). While it's simple
to add child() methods for empty tree paths to the node base classes
this does _not_ do the trick. The reason is that the base classes do
not use CRTP and can thus not return the derived node type.
This could be fixed by adding CRTP to the base classes, but this
would be a significant interface change in dune-typetree.
Thus I propose to live with the situation for now.