[concept] Cleanup isCallable()
Now isCallable(f,args...) checks if f(args...) is valid. Before,
the arguments had to be given as a Dune::TypeList which is misleading
and can even be considered a bug. Since this was not used anywhere,
this commit changes the behaviour to the expected one.

This also fixes the retun value type of isCallable() by
properly forwarding bool_constant values.