Generalize multi-index access
Implement `resolveDynamicMultiIndex()` and `resolveStaticMultiIndex()`.
While the latter simply chains `[i]` calls for all entries `i` of the
multi-index the former has to be given a termination criterion.
For the static version we require, that the containers operator[]
suppors the multi-index entries directly. For a nested multi-type
container, you can e.g. use a corresponding HybridTreePath as
multi-index.

In principle `resolveDynamicMultiIndex` superceeds
`hybridIndexAccess()` but the interface is slightly different.
However, `habridIndexAccess()` may be deprecated at a later stage.