Pass treepath to node to range map
In general using the treepath is more natural here.
While the node stores its treepath this is very likely
to be removed soon, since in any place where you need
the treepath it's available by other means (e.g. during
tree traversal).

Notice that this is an interface change! If you have
implemented your own node-to-range map, you will have
to add the additional parameter to operator().

At a later stage one may also want to drop the node argument.
However, it's not totally clear if this is needed in custom
maps.