Switch default node-to-range-map to HierarchicNodeToRangeMap
This is a non-compatible interface change. If you used a range
on a nontrivial (depth>1) tree, this will not work as before,
because the range entries are lo nonger accessed using a flat
index. (But this is very unlikely.)

As long as you only used power<...> or composite<...> in a non-nested
way, the behaviour should still be the same.