[backport to 2.6] Transform TypeTree::TreePath<...> to a container.
In 2.6 dune-typetree either provides static or dynamic
tree paths during traversal. Here we require statically sized
tree paths, which may be on the type Dune::TypeTree::TreePath<...>.
Unfortunately this does not behave like a container. As a
workaround we copy this to a Dune::TupleVector which provides
the same static information by a container-like interface.