Use static localfunctions interface if grid is single-element
The current implementation always uses the dynamic interface of
dune-localfunctions.  This is wasteful and overly complicated
in the important case of grids that know their element GeometryTypes
at compile times (like YaspGrid or ALUGrid).  With this patch,
the RaviartThomasBasis now checks at compile-time whether the
grid supports only a single element type.  If it does, the
static dune-localfunctions interface is used instead of the
dynamic one.