[doc][manual] Add footnote on pre-basis-factories
I agree with Oliver that the technical differences of pre-bases
and pre-basis-factory should not be explained to the user. However,
the text is not entirely correct and consistent if we ignore the
difference. As a way out I added a footnote explaining the fact.

The footnote is again a little technical. But for a footnote that's
OK and without it the text is simply not correct.