[doc][manual] Properly treat to wide tables/figure
I have the impression that we cannot squeeze the to
an appropriate width. As a workaround we now place them
into an explicit \makebox[\textwidth][c]{...} which has
two effects: On the one hand we pretend that the tex-box
is not to wide avoiding the latex warning. On the otherhand
the content is properly centered overlapping on both sides,
while a center-environment does not work here an makes the
content overflow to the right only.