Precompute dofs-per-entity information if order is set at run-time
The LagrangeBasis implementation needs the number of degrees of freedom
per GeometryType internally.  These depend only on the order of the
space.

So where should these values come from?  Without having done
any measurements, I conjecture that the most efficient solution
depends on whether the order is given at run-time or at
compile-time:
* with a compile-time order, the values should be recomputed every
  time they are used, because the compiler can fold the entire
  expression.  In particular, such a folded expression is faster
  than a memory access.
* with a run-time order, the expression cannot be folded, and a
  memory access is quicker than recomputing the numbers again
  and again at run-time (I may be wrong here).

This patch implements the corresponding strategy: depending on whether
the polynomial order is given at run-time or at compile-time, the
numbers are now precomputed and store, or recomputed every time
they are needed.